Add some tests over the add_extension system.
Would be nice if I refactor this whole extension system, as it's limited
to modules. You cannot register callables, or objects as an extension.

If I change this, the extensions will be specified to the most detail
object. For example::

    proj = Folio(__name__, extensions=['folio.ext.mdown.instance'])

Instead of todays::

    proj = Folio(__name__, extensions=['mdown'])

This approach will be like the Jinja2 plugin system. There will be an
`Extension` class in `folio.ext.__init__` that will be a base for every
extension written.

The cons about this that there will be a lot of more code to add,
increasing the package only for extensions. But it's true that the
extensions are the core of Folio.

It has several pros, like having a more organized way of declaring them,
maybe with common functions to provide. I'd like to make of this,
something like the Flask blueprints, but I don't know how much useful is
it.